refactor(playlist): Use anyhow::Result
A bit of a controversial choice perhaps, because at library-level anyhow
obstructs some information. And so far I've been doing completely fine
using io::{Result, Error}. But, tight library code is not the aim of
this project, it is more important for me currently that code is concise
and simple.