feat(playcount): Add Playcount::merge_duplicates()
While working on plfix I debated where to put this, and I think it
belongs directly in the playcount API, because it's a useful and natural
operation to do on a playcount file. The pretend flag is maybe
overstepping a bit, but the code is cleaner this way, so whatever.