refactor(plfix): Use vectors instead of iterators
I will need to reuse playlists and playcounts due to future upcoming
functionality. This means they have to be vectors, not one-off
iterators.
Now, the functions remove_playlist_duplicates() and
merge_playcount_duplicates() ideally would intake mutable impl IntoIterator
references, but that does not seem to be possible, because IntoIterator
doesn't guarantee that all elements returned by the iterator are unique.
Each element in an iterator loop has a lifetime equal to that of the
iterator, which means that theoretically two mutable references to the
same element could exist, which is illegal.
(source: https://stackoverflow.com/a/63438431/10792539)

I'm not 100% sure that my understanding of this is correct. In
particular, Rust has to know *sometimes* that iterator returns unique
elements. Otherwise, how would iter_mut() ever work?
For now though, I'm simply changing the argument type to &mut Vec<>,
because it does the job just fine.