feat(tracksfile): Add is_modified() method
This required changing the write() signature to take &mut self, to be
able to change the internal flag back to false. I think using interior
mutability would be better for this purpose, because for all intents and
purposes writing does not modify the structure. But that's a future step
to take.