feat(plfix): Write to log_file without storing file contents first
This removes 2 data structures that were used to accumulate all
information needed for writing to a log file. There is really no good
reason to do this, as the log file can be overwritten in real time,
which simplifies the logic and reduces memory footprint.

There was a possible bug with the prior implementation as well -- I used
indexes of tracks that came from TracksFile::tracks_unique(), which has
random order. Did not verify if this was the case, but I'm pretty sure
it didn't work as intended.