refactor!: Change playcount format
The new format removes count in favor of track duration, artist, album
and title metadata. Each line is now always a single play.

This allows for preserving information about tracks which had been
played in the past but have since been deleted from the library.

It also removes the need for an external duration cache. It also fixes
inaccurate artist names and album titles due to extracting those from
the file path itself.

Future considerations: store playlist info to be able to produce reports
on the top listened to genres? But this would probably be too difficult.