fix(track): Trim paths to be relative to MUSIC_DIR
This is only done in obvious cases. Generally speaking, Track should be
a low-level structure and not do a lot of strict checking, because it
should provide flexibility for higher-level functions.

I slightly rephrased the documentation to emphasize the fact that Tracks
must be unique within a playlist. This of course is circumvented in the
case of Playcounts, which have a wrapper Entry structure, but the idea
is that Tracks are what you use when you decide which part of a playlist
you want to remove, etc.