feat(playcount)!: Add album artist to playcount format
I realized it is needed in order to be able to calculate album listening
statistics correctly. Namely, the (album_artist, album_title) pair is
the only way to uniquely identify an album without resorting to filepath
checking (which is messier & I don't want it because I may reorganize my
directory structure in the future).