feat(plrare stats): Revamp album ranking algorithm
This implements a slightly enhanced version of the idea described here:
https://github.com/randoragon/dotfiles/commit/2ffa4ec189685acfeba210eb458ae8d9cfbb5faf

The main fix from the former lua implementation is that I obtain the
number of tracks in an album by listing its directory, not by simply
taking the length in the playcount structure. The playcount structure
does not need to have all tracks belonging to an album if they haven't
been played, so that method was error-prone, especially for smaller
periods of time like a few months.