refactor: Use soxi instead of the metadata crate
I ran into issues building the metadata crate on NixOS:

error[E0616]: field `0` of struct `ffmpeg::ChannelLayout` is private
  --> /home/pcache/.local/share/cargo/registry/src/index.crates.io-6f17d22bba15001f/metadata-0.1.9/src/stream/audio.rs:83:25
   |
83 |                 &layout.0 as *const ffmpeg::ffi::AVChannelLayout,
   |                         ^ private field

For more information about this error, try `rustc --explain E0616`.
error: could not compile `metadata` (lib) due to 1 previous error

After fighting with it for a while, I realized that the metadata crate
is pulling ffmpeg and clang into dependencies just to calculate mp3
length. I looked for alternatives and tried the mp3_duration and
symphonia crates, but was unable to make either of them work.
I then turned to an even simpler solution: just call the soxi command
and parse the output myself. Not the pretties solution, but it was
simple to write and it just works.