fix(plrare stats): Use album/track paths for key purposes
I had an album with two identically named tracks (thanks, Caravan).
This prompted me to change the data structures to use paths for
identification, not names. This is because in my filesystem, the
offending files have identical names, but different paths:

    Caravan/For Girls Who Grow Plump in the Night/Surprise, Surprise.mp3
    Caravan/For Girls Who Grow Plump in the Night/Surprise, Surprise.1.mp3

THIS SHOULD BE REFACTORED
I'm writing this at 1am, I'm in a hurry. This is super ugly. I think the
type declarations should be abandoned in favor of actual structs to make
this more readable.