feat(playlist): Add push_track() method
Sometimes you already have a Track and it's dumb to break it down into a
Utf8PathBuf just for Playlist::push() to create another Track from it
anyway. That's why this was added.

I initially tried putting it into the TracksFile trait, but it doesn't
make sense for Playcount files, so it remains Playlist-specific.