feat(tui_picker): Add did_select() method
Since plfilter requires modifying some kind of global app state on each
select, I had two options: either try to pass that global state into
on_select() callbacks, or introduce did_select() on the TuiPickerState
so that it can be checked seperately from on_select() in the main
program loop. I chose the latter because it's simpler and cleaner.
Perhaps it is not the most correct conceptually, but whatever.